Fixed Molgrum's speed cheat.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2504 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/client.h
@@ -707,7 +707,7 @@ void CL_MakeActive(char *gamename);
 
 void CL_RegisterSplitCommands(void);
 void CL_InitInput (void);
-void CL_SendCmd (float frametime);
+void CL_SendCmd (double frametime);
 void CL_SendMove (usercmd_t *cmd);
 #ifdef NQPROT
 void CL_ParseTEnt (qboolean nqprot);
